% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greens3}
\alias{greens3}
\title{Strongholds of Germany's Alliance 90/The Greens}
\format{
A data frame with 39 rows and 5 variables:
\describe{
\item{city}{name of city}
\item{city_clean}{name of city standardized}
\item{year}{year of observation}
\item{election}{federal election (BTW) or European election (EP)}
\item{greens}{vote share for the Alliance 90/The Greens}
}
}
\source{
\url{https://www.regionalstatistik.de/}
}
\usage{
greens3
}
\description{
A dataset that includes for each federal and European election the top three counties where Germany's green party received most votes.
}
\keyword{datasets}
